Fixed factoring
When vertex is on x-axis, to get a, evaluate f(0), then take the
distance from the vertex and divide it by the equivalent value in
an x^2 graph. Get the remaining graphs using the x-intercepts.
When the vertex isn't on x-axis, use similar method but evaluate
one of the x-intercepts and divide the distance of that point from
the vertex by the distance of the equivalent point in an x^2 graph
from its own vertex.